beta tweaks in interaction object size
I changed the size/scale of the slider buttons, as well as increased the size of the touchpoint colliders without actually changing the sphere size. I did not change the close button size as it looked really weird with them bigger...but we can play around with this a bit later - since it is a button, there isn't a collider associated with it.